requirements.txt
flask
pdfminer.six
spacy
requests
boto3
python-dotenv
gunicorn